+++ spec/requests/api/v1/merchants_request_spec.rb
@@ -40,7 +40,7 @@ describe "Merchant API" do
             expect(item[:attributes]).to have_key(:description)
             expect(item[:attributes][:description]).to be_a(String)
             expect(item[:attributes]).to have_key(:unit_price)
-            expect(item[:attributes][:unit_price]).to be_a(Integer)
+            expect(item[:attributes][:unit_price]).to be_a(Float)
             expect(item[:attributes]).to have_key(:merchant_id)
             expect(item[:attributes][:merchant_id]).to be_a(Integer)
             expect(Merchant.find(item[:attributes][:merchant_id])).to be_a(Merchant)
